refactor(core): rename event.rs to events.rs
Rename the events module for consistency with plural naming convention.
Update all imports across crates to use the new module name.
