+++ crates/everruns-core/src/events.rs
